refactor: relocate generateNonce to frontend
