+++ src/frontend/index.ts
@@ -14,3 +14,4 @@ export * from './agents/register';
 export * from './secrets';
 export * from './setup';
 export * from './media/img';
+export * from './nonce';
